See https://st.suckless.org/patches/alpha_focus_highlight/
Patch for suckless' simple terminal (st), which allows the user to
specify two distinct opacity values / background colors; one for the
focused- and one for unfocused windows' background. This enables the
user to spot the focused window at a glance. The patch is based on the
alpha patch; i.e. is to be applied after applying the alpha patch.