See: https://st.suckless.org/patches/csi_22_23/
This patch adds support for OSC 10, 11, and 12 in the way they are
implemented in most other terminals (e.g libvte, kitty). Specifically it
differs from osc1011_12 in that it treats the background and foreground
colors as distinct from palette colours 01 and 07 in order to facilitate
the use of theme setting scripts like theme.sh which expect these
colours to be distinct.