See: https://st.suckless.org/patches/newterm/
NOTE: The orphan patch was applied.

This patch allows you to spawn a new st terminal using
Ctrl-Shift-Return. It will have the same CWD (current working directory)
as the original st instance.

The getcwd_by_pid function is inspired on the function with the same
name of dvtm.

By default the current st terminal will be the parent process of the new
terminal. This can conflict with the swallow patch for dwm and can
result in the wrong st terminal window being swallowd. The orphan
variant of this patch works around this issue by spawning the new
terminal window as an orphan instead (meaning that it will have no
parent process).